+++ example/flutter_app/android/app/src/main/kotlin2/com/siyehua/spiexample/channel/flutter2native/IPhoto.java
@@ -1,0 +1,11 @@
+package com.siyehua.spiexample.channel.flutter2native;
+import java.util.ArrayList;
+ import java.util.HashMap;
+ import org.jetbrains.annotations.NotNull;
+ import org.jetbrains.annotations.Nullable;
+ import  com.siyehua.spiexample.channel.native2flutter.*;
+ import com.siyehua.spiexample.channel.ChannelManager.Result;
+public interface IPhoto {
+ 	void aaa();
+ 
+}