+++ lib/flutter2native_parse.dart
@@ -5,46 +5,83 @@ import 'package:platforms_source_gen/platforms_source_gen.dart';
 import 'package:platforms_source_gen/type_utils.dart';
 
 import 'auto_gen_class_json.dart';
+import 'file_config.dart';
 import 'manager/manager_creater.dart';
 import 'utils/flutter_file_utils.dart';
 import 'utils/android_file_utils.dart';
 import 'utils/ios_file_utils.dart';
 
 Future<void> flutter2Native(
-  String flutterPath,
-  String packageName,
-  String androidSavePath,
-  String iosProjectPrefix,
-  String iosSavePath,
-  bool nullSafeSupport,
-) async {
-  Directory directory = Directory(flutterPath + "/flutter2native");
+  FlutterPlatformConfig flutterConfig,
+  bool nullSafeSupport, {
+  AndroidPlatformConfig? androidConfig,
+  IosPlatformConfig? iosConfig,
+}) async {
+  Directory directory =
+      Directory(flutterConfig.sourceCodePath + "/flutter2native");
   if (!directory.existsSync()) {
     return;
   }
   List<GenClassBean> list = await platforms_source_gen_init(
-      flutterPath + "/flutter2native" //you dart file path
-      );
+      flutterConfig.sourceCodePath + "/flutter2native");
+
   await autoCreateJsonParse(list, directory.path, nullSafeSupport);
   list.forEach((element) {
     element.methods.removeWhere((element) => element.name == "toJson");
   });
 
-  _genFlutterImpl(flutterPath, packageName, list, nullSafeSupport);
+  _genFlutterImpl(flutterConfig.sourceCodePath, flutterConfig.channelName, list,
+      nullSafeSupport);
 
   ////////////////////////android//////////////////////////
   ////////////////////////android//////////////////////////
   ////////////////////////android//////////////////////////
-  JavaFileUtils.genJavaCode(
-      list, packageName, androidSavePath, ".flutter2native",
-      nullSafeSupport: nullSafeSupport);
+  if (androidConfig != null) {
+    list.forEach((element) {
+      //set custom save path
+      String classPath = element.path.split(":")[1];
+      classPath = "./lib" + classPath.substring(classPath.indexOf("/"));
+      Map<int, String> lines = File(classPath).readAsLinesSync().asMap();
+      for (int index in lines.keys) {
+        if (index == 0) {
+          if (lines[index]?.contains("FileConfig") != true) {
+            break;
+          }
+        } else if (lines[index]?.startsWith("//") == true) {
+          if (lines[index]?.contains("androidSavePath") == true) {
+            element.savePath = lines[index]?.split("=")[1].trim() ?? "";
+            element.savePath +=
+                "/" + androidConfig.packageName.replaceAll(".", "/");
+            element.savePath += "/flutter2native";
+          } else if (lines[index]?.contains("channelName") == true) {
+            element.channelName = lines[index]?.split("=")[1].trim() ?? "";
+          }
+          //channel name
+        } else {
+          String path = element.savePath.replaceAll("/flutter2native", "");
+          var list = ManagerUtils.javaSaveList[path];
+          if (list == null) {
+            list = [];
+            ManagerUtils.javaSaveList[path] = list;
+          }
+          list.add(JavaInfo.create(channelName: element.channelName));
+          break;
+        }
+      }
+    });
+    JavaFileUtils.genJavaCode(list, androidConfig.packageName,
+        androidConfig.savePath, ".flutter2native",
+        nullSafeSupport: nullSafeSupport);
+  }
 
   ////////////////////////ios//////////////////////////
   ////////////////////////ios//////////////////////////
   ////////////////////////ios//////////////////////////
-  ObjcFileUtils.genObjcCode(
-      list, iosProjectPrefix, iosSavePath, ".flutter2native",
-      nullSafeSupport: nullSafeSupport);
+  if (iosConfig != null) {
+    ObjcFileUtils.genObjcCode(
+        list, iosConfig.iosProjectPrefix, iosConfig.savePath, ".flutter2native",
+        nullSafeSupport: nullSafeSupport);
+  }
 }
 
 void _genFlutterImpl(
@@ -135,7 +172,7 @@ void _genFlutterImpl(
       // print("${match.input} + match:${ match.group(0)}");
       return "_" + match.group(0)!.toLowerCase();
     });
-    print(filePreName);
+    // print(filePreName);
     if (filePreName.startsWith("_")) {
       filePreName = filePreName.replaceFirst("_", "");
     }